+++ src/Connections/IConnection.ts
@@ -1,7 +1,7 @@
 import { MatrixEvent, MatrixMessageContent } from "../MatrixEvent";
 import { IssuesOpenedEvent, IssuesEditedEvent } from "@octokit/webhooks-types";
 import { GetConnectionsResponseItem } from "../provisioning/api";
-import { Appservice, IRichReplyMetadata } from "matrix-bot-sdk";
+import { Appservice, IRichReplyMetadata, StateEvent } from "matrix-bot-sdk";
 import { BridgeConfig, BridgePermissionLevel } from "../Config/Config";
 import { UserTokenStore } from "../UserTokenStore";
 import { CommentProcessor } from "../CommentProcessor";
@@ -9,7 +9,6 @@ import { MessageSenderClient } from "../MatrixSender";
 import { IBridgeStorageProvider } from "../Stores/StorageProvider";
 import { GithubInstance } from "../Github/GithubInstance";
 import "reflect-metadata";
-import { ConnectionDeclaration, ConnectionManager } from "../ConnectionManager";
 
 export type PermissionCheckFn = (service: string, level: BridgePermissionLevel) => boolean;
 
@@ -78,6 +77,17 @@ export interface IConnection {
     toString(): string;
 }
 
+
+
+export interface ConnectionDeclaration<C extends IConnection = IConnection> {
+    EventTypes: string[];
+    ServiceCategory: string;
+    provisionConnection?: (roomId: string, userId: string, data: Record<string, unknown>, opts: ProvisionConnectionOpts) => Promise<{connection: C}>;
+    createConnectionForState: (roomId: string, state: StateEvent<Record<string, unknown>>, opts: InstantiateConnectionOpts) => C|Promise<C>
+}
+
+export const ConnectionDeclarations: Array<ConnectionDeclaration> = [];
+
 export interface InstantiateConnectionOpts {
     as: Appservice,
     config: BridgeConfig,
@@ -92,7 +102,15 @@ export interface ProvisionConnectionOpts extends InstantiateConnectionOpts {
 }
 
 
-export function Connection<T extends ConnectionDeclaration>(constructor: T) {
-    ConnectionManager.registerConnectionType(constructor);
-    return constructor;
+export function Connection<T extends ConnectionDeclaration>(connectionType: T) {
+    // Event type clashes
+    if (ConnectionDeclarations.find(
+        (existingConn) => !!connectionType.EventTypes.find(
+            (evtType) => existingConn.EventTypes.includes(evtType))
+        )
+    ) {
+        throw Error(`Provisioning connection for ${connectionType.EventTypes[0]} has a event type clash with another connection`);
+    }
+    ConnectionDeclarations.push(connectionType);
+    return connectionType;
 }
